+++ src/constants.ts
@@ -23,6 +23,7 @@ export const javaBuildTool: string = 'javaBuildTool';
 export const functionSubpathSetting: string = 'functionSubpath';
 export const showBallerinaProjectCreationSetting: string = 'showBallerinaProjectCreation';
 export const mcpProjectTypeSetting: string = 'mcpProjectType';
+export const projectTemplatesCacheExpirationHoursSetting: string = 'projectTemplates.cacheExpirationHours';
 
 export const browseItem: IAzureQuickPickItem<undefined> = { label: localize('browse', '$(file-directory) Browse...'), description: '', data: undefined };
 
@@ -35,6 +36,7 @@ export enum ProjectLanguage {
     JavaScript = 'JavaScript',
     PowerShell = 'PowerShell',
     Python = 'Python',
+    Go = 'Go',
     TypeScript = 'TypeScript',
     Ballerina = 'Ballerina',
     Custom = 'Custom',
@@ -76,6 +78,7 @@ export enum JavaBuildTool {
 }
 
 export const ballerinaTomlFileName: string = "Ballerina.toml";
+export const goModFileName: string = 'go.mod';
 export enum BallerinaBackend {
     jvm = 'jvm',
     native = 'native'
@@ -140,6 +143,11 @@ export const extInstallTaskName: string = `${func}: ${extInstallCommand}`;
 export const hostStartCommand: string = 'host start';
 export const hostStartTaskName: string = `${func}: ${hostStartCommand}`;
 export const hostStartTaskNameRegExp = new RegExp(hostStartTaskName, 'i');
+
+// flag used by func core tools to indicate to wait for the debugger to attach before starting the worker
+export const dotnetIsolatedDebugFlag: string = '--dotnet-isolated-debug';
+export const enableJsonOutputFlag: string = '--enable-json-output';
+export const jsonOutputFileFlag: string = '--json-output-file';
 
 export const packCommand: string = 'pack';
 export const buildNativeDeps: string = '--build-native-deps';